requirements.txt
openai
streamlit
fpdf2